# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wanderbound",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/albums/{aid}/media-imports/google": {
      "post": {
        "tags": [
          "media-imports"
        ],
        "summary": "Import Google Media",
        "operationId": "import_google_media",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/ImportInProgress"
                      },
                      {
                        "$ref": "#/components/schemas/ImportCompleted"
                      },
                      {
                        "$ref": "#/components/schemas/ImportFailed"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "type",
                      "mapping": {
                        "import_in_progress": "#/components/schemas/ImportInProgress",
                        "import_completed": "#/components/schemas/ImportCompleted",
                        "import_failed": "#/components/schemas/ImportFailed"
                      }
                    }
                  },
                  "title": "Response 200 Import Google Media"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/google-photos/authorize": {
      "get": {
        "tags": [
          "google-photos"
        ],
        "summary": "Authorize",
        "operationId": "authorize",
        "parameters": [
          {
            "name": "nonce",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 64,
              "title": "Nonce"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/google-photos/callback": {
      "get": {
        "tags": [
          "google-photos"
        ],
        "summary": "Google Photos Callback",
        "operationId": "google_photos_callback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/google-photos/sessions/{session_id}": {
      "get": {
        "tags": [
          "google-photos"
        ],
        "summary": "Poll Session",
        "operationId": "poll_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "google-photos"
        ],
        "summary": "Close Session",
        "operationId": "close_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 256,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/google-photos/match/{aid}": {
      "get": {
        "tags": [
          "google-photos"
        ],
        "summary": "Match Media",
        "operationId": "match_media",
        "parameters": [
          {
            "name": "aid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Aid"
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PickerSessionId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    },
                    "event": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "retry": {
                      "type": "integer",
                      "minimum": 0
                    }
                  }
                },
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/MatchInProgress"
                      },
                      {
                        "$ref": "#/components/schemas/DownloadInProgress"
                      },
                      {
                        "$ref": "#/components/schemas/MatchCompleted"
                      },
                      {
                        "$ref": "#/components/schemas/UpgradeCompleted"
                      },
                      {
                        "$ref": "#/components/schemas/UpgradeFailed"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "type",
                      "mapping": {
                        "match_in_progress": "#/components/schemas/MatchInProgress",
                        "download_in_progress": "#/components/schemas/DownloadInProgress",
                        "match_completed": "#/components/schemas/MatchCompleted",
                        "upgrade_completed": "#/components/schemas/UpgradeCompleted",
                        "upgrade_failed": "#/components/schemas/UpgradeFailed"
                      }
                    }
                  },
                  "title": "Response 200 Match Media"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GoogleImportRequest": {
        "properties": {
          "context": {
            "type": "string",
            "enum": [
              "step",
              "cover"
            ],
            "title": "Context"
          },
          "step_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Id"
          },
          "session_id": {
            "$ref": "#/components/schemas/PickerSessionId"
          }
        },
        "type": "object",
        "required": [
          "context",
          "session_id"
        ],
        "title": "GoogleImportRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportCompleted": {
        "properties": {
          "type": {
            "type": "string",
            "const": "import_completed",
            "title": "Type",
            "default": "import_completed"
          },
          "names": {
            "items": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\\.(jpg|mp4)$"
            },
            "type": "array",
            "title": "Names"
          }
        },
        "type": "object",
        "required": [
          "names"
        ],
        "title": "ImportCompleted"
      },
      "ImportFailed": {
        "properties": {
          "type": {
            "type": "string",
            "const": "import_failed",
            "title": "Type",
            "default": "import_failed"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ImportFailed"
      },
      "ImportInProgress": {
        "properties": {
          "type": {
            "type": "string",
            "const": "import_in_progress",
            "title": "Type",
            "default": "import_in_progress"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "done",
          "total"
        ],
        "title": "ImportInProgress"
      },
...
      "MatchInProgress": {
        "properties": {
          "type": {
            "type": "string",
            "const": "match_in_progress",
            "title": "Type",
            "default": "match_in_progress"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "done",
          "total"
        ],
        "title": "MatchInProgress"
      },
...
      "PickerSessionId": {
        "type": "string",
        "maxLength": 256,
        "pattern": "^[A-Za-z0-9._-]+$"
      },
...
      "SessionStatusResponse": {
        "properties": {
          "ready": {
            "type": "boolean",
            "title": "Ready"
          }
        },
        "type": "object",
        "required": [
          "ready"
        ],
        "title": "SessionStatusResponse"
      },
...
    }
  }
}